requirements.txt
GitPython
python-telegram-bot
pyTelegramBotAPI
python-dotenv